Spotify's ID search
given an input artist name (it can include a spacebar) the code will give you:
a. the artist-id
b. the tracks-id of the 10 more popular songs of the given artist. 

so, my main job here is finally done :) 

I would ask you to test it, 
HOW TO Test: 
1.- input artist name 
2.- copy the artist-id and paste it after this 
https://play.spotify.com/artist/          <--there
you should arrive to the artist page in spotify
3.- choose any element in the tracks-id array and paste it after this:
https://play.spotify.com/track/       <--here
you should arrive at the ALBUM spotify site and the selected song shoud start playing 

it may take a while, cuz the spotify web player is sometimes slow at loading the sites. 

thx