PROBLEM SOLUTION
Problem number 1 
selected artist is not in spotify, it will search for another LocalMusician in the GLOBALLY saved array.
Problem Number 2 
Selected artist does exist in spotify, BUT doesnt have any songs available... it will find a new LocalMusician in the global array.

i saved town_t as global, that should be helpfull 

message to Phillip: can you take the location search out of the change song function ? so that it just asks your location once at the beginning of the discover

message to Ferran : can you add the follow artist button and an option of giving the city by text and not location ?