FOLLOW BUTTON + CITY LOCATIONS FIXED
so it's pretty much finished.
a lot of if cases through all the buttons but it works pretty nice and smooth. 
THERE'S IS STILL A PROBLEM I CAN'T FIX :
when the city given in text input has to words, for example "new york" MusicBrainz gives us so all artist that are named "YORK" it's pretty stupid, and i tiny error, but it sucks...